+++ R/fitVARMxID-fit-dt-var-mx-id.R
@@ -488,6 +488,8 @@
 #' coef(fit)
 #' vcov(fit)
 #' converged(fit)
+#' confint(fit)
+#' plot(fit)
 #'
 #' # Fit the model---------------------------------------------------------
 #' # center = FALSE